Added support for 64bit impls

+++ riscv/insts/integer/base/i/csrrw.kt
@@ -20,8 +20,8 @@ val csrrw = Instruction(
         parser = CSRTypeParser,
         impl16 = NoImplementation,
         impl32 = RawImplementation { mcode, sim ->
-            val vrs1 = sim.getReg(mcode[InstructionField.RS1].toInt())
-            val vcsr = sim.getReg(32).toInt()
+            val vrs1 = sim.getReg(mcode[InstructionField.RS1].toInt()).toLong()
+            val vcsr = sim.getReg(32).toLong()
             sim.setReg(mcode[InstructionField.RD].toInt(), vcsr)
             sim.setReg(32, vrs1)
             sim.incrementPC(mcode.length)
